[OpenMP] Fixed an issue in the test case parallel_offloading_map
There is a non-conforming use of variable-sized array in the test case `parallel_offloading_map.c`. This patch fixed it.

Reviewed By: protze.joachim

Differential Revision: https://reviews.llvm.org/D90642

cherry-pick: f5eebc25cc37332713888116beb4c2d9b09107ff
https://github.com/llvm/llvm-project/commit/f5eebc25cc37332713888116beb4c2d9b09107ff
